Added method for status changing

+++ src/main/java/com/swedbank/itacademy/teamcProject/repository/LoansRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by benas on 17.2.24.
  */
-public interface LoansRepository extends CrudRepository<Loans, Long> {
+public interface LoansRepository extends CrudRepository<Loans, Integer> {
 
 }
